consumer/requirements.txt
python-qpid-proton
argparse
pyyaml
jsonpath-rw
requests